requirements.txt
streamlit==1.24.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
bcrypt==4.0.1